refactor(frontend): split layout into component files
